fix: fix minor bugs after refactor
