+++ src/database/client.py
@@ -8,7 +8,6 @@ from src.database.accounts.cash.models import CashAccount
 from src.database.accounts.cash.table import CashAccountsTable
 from src.database.accounts.credit.models import CreditAccount
 from src.database.accounts.credit.table import CreditAccountsTable
-
 from src.database.accounts.models import Account
 from src.database.models import AccountTransaction
 from src.database.plaid_items.model import PlaidItem
@@ -183,7 +182,7 @@ class WalterDB:
         """
         # create account id to accounts dict
         accounts = {}
-        for account in self.get_cash_accounts(user_id):
+        for account in self.get_accounts(user_id):
             accounts[account.get_account_id()] = account
 
         # get transactions owned by user over given date range
@@ -237,7 +236,18 @@ class WalterDB:
     ############
 
     def get_accounts(self, user_id: str) -> List[Account]:
-        pass
+        cash_accounts = self.get_cash_accounts(user_id)
+        credit_accounts = self.get_credit_accounts(user_id)
+        return cash_accounts + credit_accounts
+
+    def get_account(self, user_id: str, account_id: str) -> Optional[Account]:
+        cash_account = self.get_cash_account(user_id, account_id)
+        if cash_account:
+            return cash_account
+        credit_account = self.get_credit_account(user_id, account_id)
+        if credit_account:
+            return credit_account
+        return None
 
     #################
     # CASH ACCOUNTS #
